requirements.txt
oracledb==3.0.0
python-dotenv==1.1.0
Flask==3.1.0
flasgger==0.9.7.1
pyodbc==5.2.0
psycopg2
tabulate==0.9.0
gunicorn==23.0.0